* Artikel konnte gespeichert/ gelöscht werden, wenn in Bearbeitung durch anderen Benutzer * gelöschte Artikel sind sichtbar, wenn Benutzer eingeloggt

+++ inc/controller/action/articles/articleedit.php
@@ -224,6 +224,10 @@
                 return true;
             }
             
+            if ($this->article->isInEdit()) {
+                return false;
+            }
+            
             if ($this->article->delete()) {
                 $this->redirect('articles/listall');
                 return true;
@@ -244,7 +248,7 @@
 
             $allTimer = time();
             
-            if ($this->buttonClicked('articleSave') && !$this->showRevision && $this->checkPageToken) {
+            if ($this->buttonClicked('articleSave') && !$this->showRevision && $this->checkPageToken && !$this->article->isInEdit()) {
 
                 $this->article->prepareRevision();
